Revert "make support for unsigned long and double"
This reverts commit 70da3bd5

+++ python/MIDASWrapper.cpp
@@ -14,17 +14,17 @@ PYBIND11_MODULE(MIDAS, m) {
     m.doc() = "MIDAS wrapper";
 
     py::class_<MIDAS::NormalCore>(m, "MIDAS")
-            .def(py::init<unsigned long, unsigned long>(), py::arg("num_row"), py::arg("num_col"))
+            .def(py::init<int, int>(), py::arg("num_row"), py::arg("num_col"))
             .def("add_edge", &MIDAS::NormalCore::operator(), py::arg("source"), py::arg("destination"),
                  py::arg("timestamp"));
 
     py::class_<MIDAS::RelationalCore>(m, "MIDASR")
-            .def(py::init<unsigned long, unsigned long, float>(), py::arg("num_row"), py::arg("num_col"), py::arg("factor") = 0.5)
+            .def(py::init<int, int, float>(), py::arg("num_row"), py::arg("num_col"), py::arg("factor") = 0.5)
             .def("add_edge", &MIDAS::RelationalCore::operator(), py::arg("source"), py::arg("destination"),
                  py::arg("timestamp"));
 
     py::class_<MIDAS::FilteringCore>(m, "MIDASF")
-            .def(py::init<unsigned long, unsigned long, double, double>(), py::arg("num_row"), py::arg("num_col"), py::arg("threshold"),
+            .def(py::init<int, int, float, float>(), py::arg("num_row"), py::arg("num_col"), py::arg("threshold"),
                  py::arg("factor") = 0.5)
             .def("add_edge", &MIDAS::FilteringCore::operator(), py::arg("source"), py::arg("destination"),
                  py::arg("timestamp"));
